+++ Source/Bifrost.FluentValidation/Commands/ICommandInputValidator.cs
@@ -17,6 +17,7 @@
 //
 #endregion
 using Bifrost.Commands;
+using Bifrost.Conventions;
 using Bifrost.Validation;
 using FluentValidation;
 
@@ -25,7 +26,11 @@ namespace Bifrost.FluentValidation.Commands
     /// <summary>
     /// Defines a marker interface for input level validator for a <see cref="ICommand"/>
     /// </summary>
-    public interface ICommandInputValidator : ICanValidate, IValidator
+    /// <remarks>
+    /// Types inheriting from this interface and also <see cref="ICanValidate{T}"/> will be automatically registered.
+    /// You most likely want to subclass <see cref="CommandInputValidator{T}"/>.
+    /// </remarks>
+    public interface ICommandInputValidator : ICanValidate, IValidator, IConvention
     {
     }
 }