+++ Source/Bifrost/Commands/ICommand.cs
@@ -17,13 +17,19 @@
 //
 #endregion
 using System;
+using Bifrost.Conventions;
 
 namespace Bifrost.Commands
 {
     /// <summary>
-    /// Defines the basic command
+    /// Defines the basic command.
     /// </summary>
-    public partial interface ICommand
+    /// <remarks>
+    /// Types inheriting from this interface will be picked up proxy generation, deserialized and dispatched to the
+    /// correct instance of <see cref="IHandleCommands"/>.
+    /// You most likely want to subclass <see cref="Command"/>.
+    /// </remarks>
+    public interface ICommand : IConvention
     {
         /// <summary>
         /// Gets or sets the Id of the object the command will apply to