+++ Source/Bifrost/Commands/IHandleCommands.cs
@@ -16,10 +16,12 @@
 // limitations under the License.
 //
 #endregion
+using Bifrost.Conventions;
+
 namespace Bifrost.Commands
 {
     /// <summary>
-    /// Marker interface for command handlers
+    /// Marker interface for command handlers.
     /// </summary>
     /// <remarks>
     /// A command handler must then implement a Handle method that takes the
@@ -27,9 +29,9 @@ namespace Bifrost.Commands
     ///
     /// The system will automatically detect your command handlers and methods
     /// and call it automatically when a <see cref="ICommand">command</see>
-    /// comes into the system
+    /// comes into the system.
     /// </remarks>
-    public interface IHandleCommands
+    public interface IHandleCommands : IConvention
     {
     }
 }