+++ Source/Bifrost/Events/IEvent.cs
@@ -17,13 +17,18 @@
 //
 #endregion
 using System;
+using Bifrost.Conventions;
 
 namespace Bifrost.Events
 {
     /// <summary>
-    /// Defines the basics of an event
+    /// Defines the basics of an event.
     /// </summary>
-    public interface IEvent
+    /// <remarks>
+    /// Types inheriting from this interface can be used in event sourcing and will be picked up by the event migration system.
+    /// You most likely want to subclass <see cref="Event"/>.
+    /// </remarks>
+    public interface IEvent : IConvention
     {
         /// <summary>
         /// Gets or sets the id of the event