+++ Source/Bifrost/Exceptions/IExceptionSubscriber.cs
@@ -1,11 +1,16 @@
 ﻿using System;
+using Bifrost.Conventions;
 
 namespace Bifrost.Exceptions
 {
     /// <summary>
-    /// Implement this interface to subscribe to exceptions
+    /// Implement this interface to subscribe to exceptions.
     /// </summary>
-    public interface IExceptionSubscriber
+    /// <remarks>
+    /// Types inheriting from this interface will be automatically registered.
+    /// An application can implement any number of these conventions.
+    /// </remarks>
+    public interface IExceptionSubscriber : IConvention
     {
         /// <summary>
         /// Handle the exception.