+++ Source/Bifrost/Execution/ICanSpecifyAssemblies.cs
@@ -17,6 +17,7 @@
 //
 #endregion
 using Bifrost.Configuration.Assemblies;
+using Bifrost.Conventions;
 
 namespace Bifrost.Execution
 {
@@ -24,13 +25,11 @@ namespace Bifrost.Execution
     /// Specifies what assemblies to include or not.
     /// </summary>
     /// <remarks>
-    /// Typically used by implementations of <see cref="IAssemblies"/> to
-    /// get the correct assemblies located for things like implementations of
-    /// <see cref="ITypeDiscoverer"/> which relies on knowing about assemblies
-    /// to be able to discover types.
+    /// Types inheriting from this interface will be automatically registered.
+    /// An application can implement any number of these conventions.
     /// Implementing classes must have a default constructor.
     /// </remarks>
-    public interface ICanSpecifyAssemblies
+    public interface ICanSpecifyAssemblies : IConvention
     {
         /// <summary>
         /// Method that gets called for specifying which assemblies to include or not.