+++ Source/Bifrost/Read/FetchingSecurityManager.cs
@@ -16,7 +16,6 @@
 // limitations under the License.
 //
 #endregion
-using System;
 using Bifrost.Security;
 
 namespace Bifrost.Read
@@ -26,7 +25,7 @@ namespace Bifrost.Read
     /// </summary>
     public class FetchingSecurityManager : IFetchingSecurityManager
     {
-        ISecurityManager _securityManager;
+        readonly ISecurityManager _securityManager;
 
         /// <summary>
         /// Initializes a new instance of <see cref="FetchingSecurityManager"/>
@@ -46,11 +45,6 @@ namespace Bifrost.Read
         public AuthorizationResult Authorize(IQuery query)
         {
             return _securityManager.Authorize<Fetching>(query);
-        }
-
-        public AuthorizationResult Authorize<T>(IQueryFor<T> queryFor) where T : IReadModel
-        {
-            return _securityManager.Authorize<Fetching>(queryFor);
         }
 #pragma warning restore 1591 // Xml Comments
     }